Refactor code from function decorator base to decorator base.
